+++ sys/GraphicsP.h
@@ -2,7 +2,7 @@
 #define _GraphicsP_h_
 /* GraphicsP.h
  *
- * Copyright (C) 1992-2010 Paul Boersma
+ * Copyright (C) 1992-2011 Paul Boersma
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,7 +20,7 @@
  */
 
 /*
- * pb 2010/07/30
+ * pb 2011/01/05
  */
 
 #ifndef _Graphics_h_
@@ -55,13 +55,13 @@ typedef struct {
 		/* Also used as a boolean. */ \
 	int resolution; \
 		/* Dots per inch. */ \
-	short x1DCmin, x2DCmax, y1DCmin, y2DCmax; \
+	long x1DCmin, x2DCmax, y1DCmin, y2DCmax; \
 		/* Maximum dimensions of the output device. */ \
 		/* x1DCmin < x2DCmax; y1DCmin < y2DCmax; */ \
 		/* The point (x1DCmin, y1DCmin) can be either in the top left */ \
 		/* or in the bottom left, depending on the yIsZeroAtTheTop flag. */ \
 		/* Device variables. */ \
-	short x1DC, x2DC, y1DC, y2DC; \
+	long x1DC, x2DC, y1DC, y2DC; \
 		/* Current dimensions of the output device, or: */ \
 		/* device coordinates of the viewport rectangle. */ \
 		/* x1DCmin <= x1DC < x2DC <= x2DCmax; */ \
@@ -210,7 +210,7 @@ enum opcode { SET_VIEWPORT = 101, SET_INNER, UNSET_INNER, SET_WINDOW,
 };
 
 void _Graphics_text_init (I);
-void _Graphics_fillRectangle (I, short x1DC, short x2DC, short y1DC, short y2DC);
+void _Graphics_fillRectangle (I, long x1DC, long x2DC, long y1DC, long y2DC);
 void _Graphics_setColour (I, Graphics_Colour colour);
 void _Graphics_setGrey (I, double grey);
 void _Graphics_colour_init (I);
